+++ flutter_main_page/lib/main.dart
@@ -1,12 +1,15 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_main_page/login_page.dart';
-import 'package:flutter_main_page/main_page.dart';
+import 'package:flutter_main_page/pages/loginPage/login_page.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+late SharedPreferences prefs; //안드로이드만 가능함.
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
+  prefs = await SharedPreferences.getInstance(); // 안드로이드만 가능함.
 
   runApp(MyApp());
   Fluttertoast.showToast(
@@ -18,8 +21,7 @@ void main() async {
 
 // ignore: must_be_immutable
 class MyApp extends StatelessWidget {
-  var isAutoLogin = false;
-  MyApp({Key? key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
