+++ flutter_main_page/pubspec.yaml
@@ -42,6 +42,7 @@ dependencies:
   firebase_auth: ^3.3.20
   fluttertoast: ^8.0.9
   flutter_secure_storage: ^5.0.2
+  shared_preferences: ^2.0.15
 
 dev_dependencies:
   flutter_test:
